requirements.txt
sentence-transformers>=2.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
